feat: break down the main file into core.
